+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fix deprecation warning for `re.sub`
 - Add `unrar` dependency
-- `create-stac-collection`: Set temporaly property if it is none to default
+- `create-stac-collection`:
+  - Set temporal property parameter from none to the actual configured default
+  - Support for GeoJSON input
+
 
 ## [v0.2.4] - 2025-08-27