+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v0.2.5] - 2025-08-29
+
 - Fix deprecation warning for `re.sub`
 - Add `unrar` dependency
 - `create-stac-collection`:
@@ -46,7 +48,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - First release based on vecorel CLI 0.1.0
 
-[Unreleased]: <https://github.com/vecorel/cli/compare/v0.2.3...main>
+[Unreleased]: <https://github.com/vecorel/cli/compare/v0.2.5...main>
+[v0.2.5]: <https://github.com/vecorel/cli/compare/v0.2.4...v0.2.5>
+[v0.2.4]: <https://github.com/vecorel/cli/compare/v0.2.3...v0.2.4>
 [v0.2.3]: <https://github.com/vecorel/cli/compare/v0.2.2...v0.2.3>
 [v0.2.2]: <https://github.com/vecorel/cli/compare/v0.2.1...v0.2.2>
 [v0.2.1]: <https://github.com/vecorel/cli/compare/v0.2.0...v0.2.1>